arkitscenes/ca1m: review fixes — one download engine, neutral recording module
From the code-quality review of this stack:

- Delete the bespoke _download curl wrapper and the HEAD/Content-Length
  tar-cache state machine; use the package's canonical download_file
  (atomic .part+rename, curl -C - resume, 5 retries). A tar at the final
  path is now complete by construction, so the cache-poisoning class of
  bugs cannot occur. Failures raise at the CA-1M boundary.
- Move atomic_recording out of the 558-line raw-ingest orchestrator into
  ingest/recording.py; ca1m no longer imports ingest.cli, and
  ingest/__init__.py no longer eagerly imports the orchestrator.
- Name the optional layer dirs via GT_POSES_LAYER/GT_DEPTH_LAYER instead
  of repeated string literals.
