+++ packages/arkitscenes-download/arkitscenes_download/ingest/__init__.py
@@ -1,5 +1,1 @@
-"""Convert a raw ARKitScenes sequence to a Rerun recording."""
-
-from arkitscenes_download.ingest.cli import Config, ingest_sequence
-
-__all__: list[str] = ["Config", "ingest_sequence"]
+"""Convert a raw ARKitScenes sequence to layered Rerun recordings."""
